refactor: melhora configuração do user_repository
